Todos per user

+++ server/src/todos/todos.controller.ts
@@ -1,4 +1,13 @@
-import { Controller, Get, Post, Delete, Body, Param } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Delete,
+  Body,
+  Param,
+  Req,
+} from '@nestjs/common';
+import { AuthRequest } from '../auth/interfaces';
 import { TodosService } from './todos.service';
 
 @Controller('todos')
@@ -6,17 +15,24 @@ export class TodosController {
   constructor(private readonly todosService: TodosService) {}
 
   @Get()
-  async findAll() {
-    return await this.todosService.findAll();
+  async findAll(@Req() request: AuthRequest) {
+    const userId = request.user.sub;
+    return this.todosService.findByUserId(userId);
   }
 
   @Post()
-  async create(@Body('title') title: string) {
-    return await this.todosService.create(title);
+  async create(@Req() request: AuthRequest, @Body('title') title: string) {
+    const userId = request.user.sub;
+    return await this.todosService.createForUser(title, userId);
   }
 
   @Delete(':id')
-  async remove(@Param('id') id: string) {
+  async remove(@Req() request: AuthRequest, @Param('id') id: string) {
+    const userId = request.user.sub;
+    const todo = await this.todosService.findOne(+id);
+    if (todo.user.id !== userId) {
+      throw new Error('Unauthorized');
+    }
     return await this.todosService.remove(+id);
   }
 }
